requirements.txt
aws-cdk-lib==2.72.0
constructs>=10.0.0,<11.0.0

streamlit
requests
jinja2
matplotlib
boto3
sagemaker
